Pull out context manager to top level.
